bff/requirements.txt
fastapi>=0.125.0,<1.0.0
uvicorn[standard]==0.34.0
supabase==2.11.0
pydantic>=2.11.1,<3.0.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
google-adk>=1.24.0
pytest==8.3.4
pytest-asyncio==0.25.0
